+++ powerplantmatching/cleaning.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Copyright 2016-2018 Fabian Hofmann (FIAS), Jonas Hoersch (KIT, IAI) and
 # Fabian Gotzens (FZJ, IEK-STE)
 
@@ -21,8 +20,6 @@ Functions for vertically cleaning a dataset.
 from __future__ import absolute_import, print_function
 
 import logging
-import os
-import re
 
 import networkx as nx
 import numpy as np
@@ -161,14 +158,13 @@ def gather_and_replace(df, mapping):
     for key, pattern in mapping.items():
         if not pattern:
             # if pattern is not given, fall back to case-insensitive key
-            pattern = r"(?i)\b%s\b" % key
+            pattern = rf"(?i)\b{key}\b"
         elif isinstance(pattern, list):
             # if pattern is a list, concat all entries in a case-insensitive regex
             pattern = r"(?i)" + "|".join([rf"\b{p}\b" for p in pattern])
         elif not isinstance(pattern, str):
             raise ValueError(f"Pattern must be string or list, not {type(pattern)}")
-        func = lambda ds: ds.str.contains(pattern)
-        where = df.astype(str).apply(func).any(axis=1)
+        where = df.astype(str).apply(lambda ds: ds.str.contains(pattern)).any(axis=1)
         res = res.where(~where, key)
     return res
 
@@ -411,7 +407,7 @@ def aggregate_units(
         for arg in used_deprecated_args:
             kwargs.pop(arg)
         msg = "The following arguments were deprecated and are being ignored: "
-        logger.warn(msg + f"{used_deprecated_args}")
+        logger.warning(msg + f"{used_deprecated_args}")
 
     df = get_obj_if_Acc(df)
 
@@ -453,7 +449,7 @@ def aggregate_units(
 
     # Downcasting in replace is deprecated
     with pd.option_context("future.no_silent_downcasting", True):
-        df[str_cols] = df[str_cols].replace("", np.nan).infer_objects(copy=False)
+        df[str_cols] = df[str_cols].replace("", pd.NA).infer_objects(copy=False)
 
     df = (
         df.assign(